+++ pkg/artifactory/datasource/repository/local/datasource_artifactory_local_cargo_repository.go
@@ -2,6 +2,7 @@ package local
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/jfrog/terraform-provider-artifactory/v7/pkg/artifactory/datasource/repository"
 	"github.com/jfrog/terraform-provider-artifactory/v7/pkg/artifactory/resource/repository/local"
 	"github.com/jfrog/terraform-provider-shared/packer"
 )
@@ -11,14 +12,14 @@ func DataSourceArtifactoryLocalCargoRepository() *schema.Resource {
 		return &local.CargoLocalRepoParams{
 			RepositoryBaseParams: local.RepositoryBaseParams{
 				PackageType: "cargo",
-				Rclass:      "local",
+				Rclass:      rclass,
 			},
 		}, nil
 	}
 
 	return &schema.Resource{
 		Schema:      local.CargoLocalSchema,
-		ReadContext: MkRepoReadDataSource(packer.Default(local.CargoLocalSchema), constructor),
+		ReadContext: repository.MkRepoReadDataSource(packer.Default(local.CargoLocalSchema), constructor),
 		Description: "Data source for local cargo repository",
 	}
 }